Rename ZulipNative to ZulipMobile, fix issues with latest packages
