+++ android/app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
 
-    <string name="app_name">ZulipNative</string>
+    <string name="app_name">ZulipMobile</string>
 </resources>
