+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'ZulipNative'
+rootProject.name = 'ZulipMobile'
 
 include ':app'
 include ':react-native-vector-icons'
