+++ ios/ZulipMobile/AnchoredScrollView.m
