+++ ios/ZulipMobile/Images.xcassets/AppIcon.appiconset/Contents.json
