+++ ios/ZulipMobile/Images.xcassets/AppIcon.appiconset/Icon-40@2x.png
