+++ ios/ZulipMobile/Images.xcassets/AppIcon.appiconset/Icon-60@3x.png
