+++ ios/ZulipMobile/Images.xcassets/AppIcon.appiconset/Icon-Small-50.png
