+++ ios/ZulipMobile/Images.xcassets/AppIcon.appiconset/Icon-Small@2x.png
