+++ ios/ZulipMobileTests/Info.plist
