+++ package.json
@@ -1,10 +1,16 @@
 {
-  "name": "ZulipNative",
-  "version": "0.0.1",
+  "name": "ZulipMobile",
+  "version": "0.9.0",
   "private": true,
   "scripts": {
     "start": "react-native start",
-    "ios": "react-native run-ios",
+    "ios-min": "react-native run-ios --simulator 'iPhone 5'",
+    "ios-max": "react-native run-ios --simulator 'iPhone 7'",
+    "ios-device": "react-native run-ios --device 'Max's iPhone'",
+    "android-min": "react-native run-android --simulator '4.0'",
+    "android-max": "react-native run-android --simulator '7.0'",
+    "android-max": "react-native run-android --simulator '7.0'",
+    "android-release": "react-native run-android --configuration 'Release'",
     "test:lint": "eslint src/",
     "test:flow": "node_modules/.bin/flow check",
     "test:unit": "jest",
